Move all utils to utils dir
