+++ scargo/commands/check.py
@@ -12,10 +12,10 @@ from itertools import chain
 from pathlib import Path
 from typing import Iterable, List, NamedTuple, Optional, Sequence, Type
 
-from scargo.clang_utils import get_comment_lines
 from scargo.config import CheckConfig, Config, TodoCheckConfig
 from scargo.config_utils import prepare_config
 from scargo.logger import get_logger
+from scargo.utils.clang_utils import get_comment_lines
 
 logger = get_logger()
