+++ scargo/commands/debug.py
@@ -13,9 +13,9 @@ from typing import List, Optional
 
 from scargo.config import CHIP_DEFAULTS, Config, ScargoTarget
 from scargo.config_utils import get_target_or_default, prepare_config
-from scargo.docker_utils import run_scargo_again_in_docker
 from scargo.logger import get_logger
-from scargo.sys_utils import find_program_path
+from scargo.utils.docker_utils import run_scargo_again_in_docker
+from scargo.utils.sys_utils import find_program_path
 
 logger = get_logger()
